Migrate basic bot functionality
Migrate the bot with the ping and info commands working correctly.

Several file structure changes have been made to improve the long-term development process.
The bot is now constructed in a subclass of CommandClient and exposed as a singleton.
"Modules" have been replaced with controllers to be more in line with the traditional MVC framework.
Loaders are now used to construct controller instances and register the commands.