Fix incorrect usage of redislock
The previous implementation of calendar locking was done with an
assumption that redislock would make different locks from a single
redislock class instance by using the key supplied to the acquire()
method calls. In reality, the keys only serve as an identifier
for that specific lock when it is acquired, and in the bot's case,
one lock instance made by calling createLock can only be used by
one guild at a time.

This commit changes the CalendarLock class to create a new lock
each time a static acquire method call is made to the class, and
returns the locked lock for releasing later. The lock's usage in
the calendar controller and event scheduler have been adjusted
accordingly.

Closes #4.